requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
libsnmp==2.0.5
MarkupSafe==1.1.1
pexpect==4.8.0
pkg-resources==0.0.0
ply==3.11
ptyprocess==0.6.0
pyasn1==0.4.8
pycrypto==2.6.1
pycryptodomex==3.9.7
pysmi==0.3.4
pysnmp==4.4.12
pysnmp-mibs==0.1.6
Werkzeug==1.0.0
